consolidate SBP compliance into FAD controller
