+++ AIS/AIS/Views/FAD/AuditValidation.cshtml
@@ -5,7 +5,7 @@
 }
 <h4>Audit Validation</h4>
 <p>The Internal Audit team reviews the observation and may validate it, refer it back, or raise a query.</p>
-<form asp-action="AuditValidation" asp-controller="SBPCompliance" method="post">
+<form asp-action="AuditValidation" asp-controller="FAD" method="post">
     <input type="hidden" asp-for="ObservationId" />
     <div class="form-group">
         <label asp-for="Remarks"></label>
